Fixing issue #5 and #6
Re-worked the capturing of bystander so that an option to capture a
bystander could be added to the card actions.  Added the option to draw
from the villain deck right from the villain deck so that if a card
action causes you to draw from the villain deck you can do so and it
will run through the automation of what should happen during the villain
draw like it does when you start your turn.